refactor: 파일 위치 변경
각 파일을 domain, enums 디렉토리 내로 이동합니다.
