+++ src/main/java/mana/doodleking/domain/user/enums/UserRole.java
@@ -1,4 +1,4 @@
-package mana.doodleking.domain.user;
+package mana.doodleking.domain.user.enums;
 
 public enum UserRole {
     LEADER, MEMBER
