+++ src/main/java/mana/doodleking/domain/user/enums/UserState.java
@@ -1,4 +1,4 @@
-package mana.doodleking.domain.user;
+package mana.doodleking.domain.user.enums;
 
 public enum UserState {
     READY, NOT_READY
